Fix go installation, add nvim installation from PPA
- Installing go as a snap might make our go installation weirdness
easier to manage.  In addition to reducing the number of steps we take
to install go, it creates a symlink so that we don't have to keep
updating our bashit custom paths to find what we install.

- I've finally figured out why we're installing neovim from PPA: it's
the only way to get versions above 0.2.x, which is required by our
plugins.  I added that back in.

- Fixed a bug installing the cli source to the go path.